chore: use mocha global hook to register before/after
